newest version

+++ annotate_texts.py
@@ -11,7 +11,7 @@ import re
 
 start_time = time.process_time()
 total_timed_out = 0
-MAX_TIME_OUT = 1
+MAX_TIME_OUT = 30
 
 
 def annotate_text(url, input_path, output_path_text, output_path_csv, max_size):
@@ -61,19 +61,23 @@ def annotate_text(url, input_path, output_path_text, output_path_csv, max_size):
 def send_requests(text_chunks, url, output_path_text, output_path_csv, total_timed_out=total_timed_out):
     for text in text_chunks:
         timeout = 8
-
         len_words = len(text.split())
         print("Text length:", len(text), "NUM WORDS", len_words)
+        print(text)
+
+        level = os.path.basename(os.path.dirname(input_path))
 
         if len_words > 250:
             timeout = 17
         elif len_words > 150:
             timeout = 12
+        if level in {"b2", "c1", "c2"}:
+            timeout = 35
+            print("HARD LEVEL!")
 
         text = text.replace(" ", "%20").replace("\n", "%0A")
         url = url + text
         try:
-            print(timeout)
             request = requests.post(url, timeout=timeout)  # Set a max time in seconds
             if request.status_code != 200:
                 print("\nError with status code {}".format(request.status_code) + ": " + input_path + "\n")
@@ -81,23 +85,15 @@ def send_requests(text_chunks, url, output_path_text, output_path_csv, total_tim
                 data = request.json()
                 # print("NUMBER OF ANNOTATIONS in " + input_path, len(data["annotationList"]))
                 write_csv_and_txt(data, output_path_text, output_path_csv)
-                os.remove(input_path)
 
+                # if the text is very long, only delete when we are done processing all the chunks
+                if text == text_chunks[-1]:
+                    os.remove(input_path)
+            total_timed_out = 0
         # If exceptions are raised, move the input file to a different directory
         except requests.exceptions.Timeout:
             print("\nRequest timed out:", input_path + "\n")
             total_timed_out += 1
-
-            # TODO restart server if too many requests have timed out
-            # if total_timed_out >= MAX_TIME_OUT:
-            #     print("Timeout threshold exceeded. Restarting the server...")
-            #
-            #     # Trigger the bash script to restart the server
-            #     subprocess.run(["bash", "restart_server.sh"])
-            #
-            #     print("Waiting for server to restart...")
-            #     time.sleep(restart_wait_time)
-            # # get filename and level
 
             filename = os.path.basename(input_path)
             level = os.path.basename(os.path.dirname(input_path))
@@ -105,9 +101,9 @@ def send_requests(text_chunks, url, output_path_text, output_path_csv, total_tim
 
             # make directory if not exists
             os.makedirs(timeout_level_path, exist_ok=True)
-
-            # move the file
-            os.rename(input_path, os.path.join(timeout_level_path, filename))
+            if text == text_chunks[-1]:
+                # move the file
+                os.rename(input_path, os.path.join(timeout_level_path, filename))
         except requests.exceptions.RequestException as e:
             print("\n\nError occurred\n\n")
             filename = os.path.basename(input_path)
@@ -147,7 +143,7 @@ def write_csv_and_txt(data, output_path_text, output_path_csv):
 
 if __name__ == "__main__":
     url = "http://localhost:4000/extractor?text="
-    input_dir = "data/input/"
+    input_dir = "data/c2_from_30k/"
     output_dir = "data/output/"
 
     for input_path in utils.get_file_paths(input_dir):
@@ -155,7 +151,7 @@ if __name__ == "__main__":
         level = os.path.split(input_path)[0][-2:]
         output_path_text = os.path.join(output_dir, level, "text", os.path.split(input_path)[1][:-4])
         output_path_csv = os.path.join(output_dir, level, "csv", os.path.split(input_path)[1][:-4])
-        annotate_text(url, input_path, output_path_text, output_path_csv, 450)
+        annotate_text(url, input_path, output_path_text, output_path_csv, 100)
 
     print("\nTime elapsed: ", time.process_time() - start_time, "seconds\n")
